Move class to single point of usage
